refactor: update files structure
