+++ src/layouts/layouts.ts
@@ -1,0 +1,1 @@
+export { default as HeaderLayout } from "./header/HeaderLayout";
