+++ src/pages/pages.ts
@@ -1,0 +1,6 @@
+export { default as HomePage } from "./home/HomePage";
+export { default as PostsPage } from "./posts/PostsPage";
+export { default as PostPage } from "./posts/PostPage";
+export { default as NotFoundPage } from "./notfound/NotFoundPage";
+export { default as RegistrationPage } from "./register/RegistrationPage";
+export { default as LoginPage } from "./login/LoginPage";
