+++ src/pages/posts/PostPage.tsx
@@ -1,6 +1,8 @@
 import { useParams } from "react-router-dom";
 
-export const PostPage = () => {
+const PostPage = () => {
   const params = useParams<{ postId: string }>();
   return <h1>Post Page {params.postId}</h1>;
 };
+
+export default PostPage;
